Moving namespace for editors
