+++ Ana/View/ViewModelLocator.cs
@@ -4,6 +4,10 @@
     using Source.ChangeLog;
     using Source.CheatBrowser;
     using Source.DotNetExplorer;
+    using Source.Editors.HotkeyEditor;
+    using Source.Editors.ScriptEditor;
+    using Source.Editors.TextEditor;
+    using Source.Editors.ValueEditor;
     using Source.Main;
     using Source.Output;
     using Source.ProcessSelector;
@@ -20,10 +24,6 @@
     using Source.SignatureCollector;
     using Source.Snapshots;
     using Source.UserSettings;
-    using Source.Utils.HotkeyEditor;
-    using Source.Utils.ScriptEditor;
-    using Source.Utils.TextEditor;
-    using Source.Utils.ValueEditor;
 
     /// <summary>
     /// This class contains static references to all the view models in the
